requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Data validation
great-expectations>=0.17.0

# Cloud storage
google-cloud-storage>=2.10.0

# Apache Beam for batch processing
apache-beam[gcp]>=2.50.0

# HTTP requests
requests>=2.31.0

# Type hints support
typing-extensions>=4.7.0

# Logging (part of Python standard library)
# logging

# JSON handling (part of Python standard library)
# json

# OS operations (part of Python standard library)
# os 